BAU: Add credentials to publishing repository config
Previously, the package was published to a directory on Jenkins box without credentials. This change adds artifactory credentials so that the package can be pushed to a remote artifactory. This change is needed to allow the pubish task to be executed on a box which does not host Jenkins or artifactory.

Author: @adityapahuja