HUB-907: Move dropwizard-client to api configuration
The `baseUri` method on `HttpStubRule` returns a UriBuilder. This is a
type defined in a dependency.

Projects consuming this library will need to have access to that type at
compile time. Moving the dependency to the `api` configuration allows
that to happen.

The `java-library` plugin gives us the `api` configuration. Using this
means the dependencies *are* exposed to the consuming projects compile
time.

This change is being prompted as a consuming project is unable to
compile due to missing these dependencies.